ticket:1106 - scons improvements for building OmeroCpp
 * By default, only "Compiling..." and "Linking..." are printed
 * Environment variable "VERBOSE=1" prints long command lines
 * Environment varaiable "CXXFLAGS=-O3" now supported.
 * Also fixed various warnings from -Wall


git-svn-id: file:///home/svn/omero/trunk@2946 05709c45-44f0-0310-885b-81a1db45b4a6

+++ components/tools/OmeroCpp/test/integration/search.cpp
@@ -113,7 +113,7 @@ public:
  */
 #define assertResults(count, search) _assertResults(__LINE__, count, search, true) 
 #define assertAtLeastResults(count, search) _assertResults(__LINE__, count, search, false) 
-void _assertResults(int line, int count, SearchPrx search, bool exact) {
+void _assertResults(int line, unsigned int count, SearchPrx search, bool exact) {
     stringstream out;
     out << "line " << line << ":";
     if (count  > 0) {
@@ -183,7 +183,7 @@ BOOST_AUTO_TEST_CASE( IQuerySearch )
 	 */
         IObjectList list;
         list = f.query()->findAllByFullText("Image",uuid,0);
-	BOOST_CHECK_EQUAL( 1, list.size() );
+	BOOST_CHECK_EQUAL( (unsigned int) 1, list.size() );
 	
     } catch (const omero::ApiUsageException& aue) {
         cout << aue.message <<endl;
@@ -1483,11 +1483,11 @@ BOOST_AUTO_TEST_CASE( testOnlyAnnotatedWithMultiple ) {
 
     search->onlyAnnotatedWith(stringSet("TagAnnotation"));
     search->byFullText(name);
-    BOOST_CHECK_EQUAL(2, search->results().size());
+    BOOST_CHECK_EQUAL( (unsigned int) 2, search->results().size());
 
     search->onlyAnnotatedWith(stringSet("BooleanAnnotation"));
     search->byFullText(name);
-    BOOST_CHECK_EQUAL(2, search->results().size());
+    BOOST_CHECK_EQUAL( (unsigned int) 2, search->results().size());
 
     search->onlyAnnotatedWith(stringSet("BooleanAnnotation", "TagAnnotation"));
     search->byFullText(name);
